requirements.txt
# attrs==25.3.0
# babel==2.17.0
# beautifulsoup4==4.13.3
# blinker==1.9.0
# certifi==2025.1.31
# cffi==1.17.1
# charset-normalizer==3.4.1
# click==8.1.8
# colorama==0.4.6
# courlan==1.3.2
# dateparser==1.2.1
# dnspython==2.7.0
# email_validator==2.2.0
# Flask==3.1.0
# Flask-Mail==0.10.0
# Flask-SQLAlchemy==3.1.1
# greenlet==3.1.1
# gunicorn==23.0.0
# h11==0.14.0
# htmldate==1.9.3
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.5
# jusText==3.0.2
# lxml==5.3.1
# lxml_html_clean==0.4.1
# MarkupSafe==3.0.2
# outcome==1.3.0.post0
# packaging==24.2
# psycopg2-binary==2.9.10
# pycparser==2.22
# PySocks==1.7.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# python-http-client==3.3.7
# pytz==2025.1
# regex==2024.11.6
# requests==2.32.3
# selenium==4.29.0
# sendgrid==6.11.0
# six==1.17.0
# sniffio==1.3.1
# sortedcontainers==2.4.0
# soupsieve==2.6
# SQLAlchemy==2.0.38
# starkbank-ecdsa==2.2.0
# tld==0.13
# trafilatura==2.0.0
# trio==0.29.0
# trio-websocket==0.12.2
# typing_extensions==4.12.2
# tzdata==2025.1
# tzlocal==5.3
# urllib3==2.3.0
# websocket-client==1.8.0
# Werkzeug==3.1.3
# wsproto==1.2.0
# schedule==1.2.1
# webdriver-manager==4.0.2

# Core web framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1

# Database
psycopg2-binary==2.9.10

# Environment & utilities
python-dotenv==1.0.1
requests==2.32.3

# Scraping & browser automation
beautifulsoup4==4.13.3
selenium==4.29.0
webdriver-manager==4.0.2

# Scheduler for background price checks
schedule==1.2.1

# Production server (needed on most hosting platforms)
gunicorn==23.0.0